moved function to get ID data to contract shim
