+++ src/user.ts
@@ -76,14 +76,6 @@ interface PublicSdk {
     signAndSendCredential(): Promise<void>;
 }
 
-type GetAccountTonomyIDInfoResponse = {
-    account_name: Name,
-    account_type: Number,
-    account_status: Number,
-    username_hash: Checksum256,
-    password_salt: Checksum256
-};
-
 class User {
     authenticator: Authenticator;
 
@@ -138,7 +130,7 @@ class User {
     static async getAccountInfo(account: string | Name): Promise<API.v1.AccountObject> {
         if (typeof account === 'string') {
             // this is a username
-            const idData = await this.getAccountTonomyIDInfo(account)
+            const idData = await idContract.getAccountTonomyIDInfo(account)
             return await api.v1.chain.get_account(idData.account_name);
         } else {
             // use the account name directly
@@ -146,28 +138,6 @@ class User {
         }
     }
 
-    static async getAccountTonomyIDInfo(account: string | Name): Promise<GetAccountTonomyIDInfoResponse> {
-        let data;
-        if (typeof account === 'string') {
-            // this is a username
-            const accountName = sha256(account);
-
-            data = await api.v1.chain.get_table_rows({
-                code: "id.tonomy",
-                table: "accounts",
-                lower_bound: Name.from(accountName),
-                index_position: "secondary"
-            });
-        } else {
-            // use the account name directly
-            data = await api.v1.chain.get_table_rows({
-                code: "id.tonomy",
-                table: "accounts",
-                lower_bound: Name.from(account)
-            });
-        }
-        return data.rows[0];
-    }
 }
 
 export { User };
